Move NFDeployment interface to its own file
